서버 엔진 리팩토링 #2  - id -> user_pid로 수정 (자료형 int64로)  - define 값 헤더 분리  - socket 참조 수정  - acceptor는 deprecate 예정.. (딱히 역할이 없음)

+++ ChattingServer/User.cpp
@@ -5,7 +5,7 @@ User::User()
 	
 }
 
-User::User(SOCKET s, int id) :
+User::User(SOCKET s, __int64 id) :
 	_userSocket(s),
 	_userId(id)
 {
